+++ test/commands/prodly/unit/version.test.ts
@@ -1,0 +1,42 @@
+import { SfError } from '@salesforce/core';
+import { TestContext } from '@salesforce/core/testSetup';
+import { expect } from 'chai';
+import ProdlyVersion from '../../../../src/commands/prodly/version.js';
+
+describe('prodly version', () => {
+  const $$ = new TestContext();
+
+  afterEach(() => {
+    $$.restore();
+  });
+
+  it('should throw an error when target dev hub is not provided', async () => {
+    try {
+      await ProdlyVersion.run(['--target-org', 'test']);
+    } catch (error) {
+      if (error instanceof SfError) {
+        expect(error.message).to.include('No default dev hub found');
+      }
+    }
+  });
+
+  it('should throw an error when target org is not provided', async () => {
+    try {
+      await ProdlyVersion.run(['--target-dev-hub', 'test']);
+    } catch (error) {
+      if (error instanceof SfError) {
+        expect(error.message).to.include('No default environment found');
+      }
+    }
+  });
+
+  it('should throw an error when source-branch flag is not provided.', async () => {
+    try {
+      await ProdlyVersion.run(['--target-org', 'test', '--target-dev-hub', 'test']);
+    } catch (error) {
+      if (error instanceof SfError) {
+        expect(error.message).to.include('Missing required flag source-branch');
+      }
+    }
+  });
+});