+++ CHANGELOG.md
@@ -1,5 +1,52 @@
 # Changelog
 
+## [1.2.0] - 2024-09-04
+
+### Added
+
+- New command `prodly:version` added. This command allows the user to add version control to a managed environment.
+
+## [1.1.3] - 2024-09-03
+
+### Changed
+
+- Updated `@salesforce/core` to v8.5.3
+- Updated `@salesforce/sf-plugins-core` to v11.3.7
+- Updated `@oclif/plugin-command-snapshot` to v5.2.15
+- Updated `@salesforce/cli-plugins-testkit` to v5.3.28
+- Updated `eslint-plugin-sf-plugin` to v1.20.6
+- Updated `oclif` to v4.14.29
+
+## [1.1.2] - 2024-08-30
+
+### Changed
+
+- Updated `@oclif/core` to v4.0.19
+- Updated `@salesforce/core` to v8.5.1
+- Updated `@salesforce/sf-plugins-core` to v11.3.6
+- Updated `@oclif/plugin-command-snapshot` to v5.2.14
+- Updated `@salesforce/cli-plugins-testkit` to v5.3.26
+- Updated `eslint-plugin-sf-plugin` to v1.20.5
+- Updated `oclif` to v4.14.26
+
+## [1.1.1] - 2024-08-08
+
+### Fixed
+
+- Fixed manage command when connection id is null.
+
+### Changed
+
+- Updated `@oclif/core` to v4.0.17
+- Updated `@salesforce/core` to v8.2.8
+- Updated `@salesforce/sf-plugins-core` to v11.2.8
+- Updated `@oclif/plugin-command-snapshot` to v5.2.12
+- Updated `@salesforce/cli-plugins-testkit` to v5.3.23
+- Updated `@salesforce/dev-scripts` to v10.2.9
+- Updated `eslint-plugin-sf-plugin` to v1.20.3
+- Updated `oclif` to v4.14.15
+- Updated `typescript` to v5.5.4
+
 ## [1.1.0] - 2024-03-22
 
 ### Added